backend/requirements/base.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
psycopg[binary]>=3.2,<3.3
redis>=5.0,<6.0
celery>=5.4,<5.5
django-environ>=0.11,<0.12
django-filter>=24.3,<25.0
drf-spectacular>=0.27,<0.28
stripe>=11.6,<12.0
